% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normala.R
\name{normala}
\alias{normala}
\title{Normalize x}
\usage{
normala(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector of numbers that will be
normalized.}

\item{na.rm}{When na.rm is TRUE, it will be
ignored in calculation. When it is FALSE, it
will be used in calculation.}
}
\value{
A vector of numbers resulting from
x
}
\description{
This function is used to normalize x
}
\details{
This is used when comparing variables
but they have different units.
The input value must be numeric.
}
\examples{
normala(1:20)
normala(c(1:20,NA),na.rm=TRUE)
}
